More refactoring

+++ ext/semian/semset.h
@@ -42,7 +42,7 @@ key_t
 generate_sem_set_key(const char *name);
 
 void
-set_semaphore_permissions(int sem_id, int permissions);
+set_semaphore_permissions(int sem_id, long permissions);
 
 int
 get_sem_val(int sem_id, int sem_index);
@@ -57,7 +57,7 @@ void
 sem_meta_unlock(int sem_id);
 
 int
-create_semaphore(int key, int permissions, int *created);
+create_semaphore(int key, long permissions, int *created);
 
 int
 get_semaphore(int key);
